requirements.txt
fastapi
uvicorn

pymupdf
python-multipart

pytesseract
Pillow

sentence-transformers

chromadb

google-generativeai

python-dotenv
tqdm

nltk
loguru


# RAGG/
# │
# ├── app/
# │   ├── __init__.py
# │   ├── main.py                # FastAPI entrypoint
# │   ├── config.py              # Configs: API keys, paths, constants
# │   │
# │   ├── routers/               # API route definitions
# │   │   ├── __init__.py
# │   │   ├── upload.py          # Upload endpoint
# │   │   ├── query.py           # Query endpoint
# │   │
# │   ├── services/              # Core logic
# │   │   ├── __init__.py
# │   │   ├── file_loader.py     # PDF/text/OCR extraction
# │   │   ├── chunker.py         # Text chunking logic
# │   │   ├── embedder.py        # SentenceTransformer embeddings
# │   │   ├── retriever.py       # ChromaDB vector search
# │   │   ├── generator.py       # Gemini API call logic
# │   │
# │   ├── db/                    # Vector DB setup and interaction
# │   │   ├── __init__.py
# │   │   ├── chroma_client.py   # Collection init and management
# │   │
# │   ├── utils/                 # Helpers (e.g. logging, timing)
# │   │   ├── __init__.py
# │   │   ├── logger.py
# │   │   ├── cleaner.py
# │
# ├── data/                      # Uploaded and processed documents
# │   ├── uploads/
# │   ├── processed/
# │
# ├── tests/                     # Unit and integration tests
# │
# ├── requirements.txt           # Python dependencies
# ├── README.md
# ├── .env                       # API keys and environment variables


# rag_app/
# │
# ├── app/
# │   ├── main.py             # FastAPI entrypoint
# │   ├── config.py           # Loads environment variables (e.g. Gemini key)
# │   ├── file_loader.py      # Load and extract text from PDFs
# │   ├── chunker.py          # Split text into chunks
# │   ├── embedder.py         # Embed chunks using SentenceTransformers
# │   ├── chroma_client.py    # ChromaDB setup and retrieval
# │   ├── gemini_client.py    # Send prompt to Gemini and get answer
# │
# ├── data/                   # Where uploaded files live
# │   └── uploads/
# │
# ├── requirements.txt        # Dependencies
# ├── .env                    # Environment variables
# ├── README.md               # (optional) Project info